% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-palette.R
\name{assign_colors}
\alias{assign_colors}
\title{Assign colors to provided names}
\usage{
assign_colors(colors, names, filter = NULL, ...)
}
\arguments{
\item{colors}{Vector of colors}

\item{names}{Vector of names}

\item{filter}{Filter to apply to color palette when
calculating pairwise differences.
A vector can be passed to adjust based on multiple color filters.
Possible values include, "none", "deutan", "protan", and "tritan".}

\item{...}{Additional arguments to pass to `grDevices::colorRamp()`}
}
\description{
Reduce or expand color palette and assign names
}
