% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-colors.R
\name{compare_colors}
\alias{compare_colors}
\title{Compare colors in palette}
\usage{
compare_colors(
  colors,
  y = NULL,
  filter = NULL,
  method = "CIE2000",
  return_mat = FALSE,
  names = TRUE
)
}
\arguments{
\item{colors}{Vector of colors}

\item{y}{Vector of colors to compare,
if `NULL`, `colors` will be compared with itself.}

\item{filter}{Filter to apply to color palette when
calculating pairwise differences.
A vector can be passed to adjust based on multiple color filters.
Possible values include,
- "colorblind", use deutan, protan, and tritan color blindness simulation
  filters
- "deutan"
- "protan"
- "tritan"}

\item{method}{Method to use for comparing colors, can be one of:
- "euclidian"
- "CIE1976"
- "CIE94"
- "CIE2000"
- "CMC"}

\item{return_mat}{Return matrix of minimum pairwise color differences for
specified color filters.}

\item{names}{Should color names be included in output}
}
\description{
Compare colors in palette
}
