% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce-colors.R
\name{get_property}
\alias{get_property}
\title{Get color property}
\usage{
get_property(colors, property)
}
\arguments{
\item{colors}{Character vector of colors}

\item{property}{Vector of color properties to return, can include any of:
- "lightness", from LAB colorspace
- "a", from LAB colorspace
- "b", from LAB colorspace
- "hue", from HSL colorspace
- "saturation", from HSL colorspace
- "red", from RGB colorspace
- "green", from RGB colorspace
- "blue", from RGB colorspace}
}
\description{
Get color property
}
