% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-colors.R
\name{interp_colors}
\alias{interp_colors}
\title{Interpolate colors}
\usage{
interp_colors(
  colors,
  n,
  keep_original = TRUE,
  order = TRUE,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{colors}{Vector of colors}

\item{n}{Number of colors to include in final color palette}

\item{keep_original}{If `TRUE` original colors will be included in final
palette}

\item{order}{If `TRUE` colors are ordered with new colors interspersed,
if `FALSE` new colors are added to the end.}

\item{before, after}{Numeric vector indicating the position in the original
palette where new colors should be inserted.
A position must be provided for each new color added.
The default is to evenly distribute new colors throughout the original
palette.}
}
\description{
Insert additional colors into palette, this behaves similar to
`scales::colour_ramp()`, but keeps the original colors in the
final palette.
}
