% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-palette.R
\name{ramp_colors}
\alias{ramp_colors}
\title{Ramp colors}
\usage{
ramp_colors(colors, n, keep_original = TRUE, order = TRUE, ...)
}
\arguments{
\item{colors}{Vector of colors}

\item{n}{Number of colors to include in final color palette}

\item{keep_original}{If `TRUE` original colors will be included in final
palette}

\item{order}{If `TRUE` colors are ordered with new colors interspersed,
if `FALSE` new colors are added to the end.}

\item{...}{Additional arguments to pass to `grDevices::colorRamp()`}
}
\description{
Insert additional colors into palette, this behaves similar to
`grDevices::colorRampPalette()`, but keeps all of the original colors in the
final palette.
}
