% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-colors.R
\name{resize_colors}
\alias{resize_colors}
\title{Resize color palette}
\usage{
resize_colors(colors, n, order = TRUE, ...)
}
\arguments{
\item{colors}{Vector of colors}

\item{n}{Number of colors to include in final color palette}

\item{order}{If `TRUE` colors are ordered with new colors interspersed,
if `FALSE` new colors are added to the end.}

\item{...}{Additional arguments to pass to `collapse_colors()`.}
}
\description{
Reduce or expand the number of colors in palette, colors are removed with
`collapse_colors()` and added with `interp_colors()`.
}
