% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-palette.R
\name{resize_palette}
\alias{resize_palette}
\title{Resize color palette}
\usage{
resize_palette(colors, n, order = TRUE, ...)
}
\arguments{
\item{colors}{Vector of colors}

\item{n}{Number of colors to include in final color palette}

\item{order}{If `TRUE`, when colors are expanded new colors are interspersed,
if `FALSE`, new colors are added to the end.}

\item{...}{Additional arguments to pass to `collapse_colors()` or
`ramp_colors()`}
}
\description{
Reduce or expand the number of colors in palette
}
