% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-spruce.R
\name{scale_color_spruce}
\alias{scale_color_spruce}
\title{Automatically adjust color values}
\usage{
scale_color_spruce(
  ...,
  values,
  difference = 10,
  adjust = "lightness",
  colorblind = "none",
  highlight = NULL,
  range = NULL,
  maxit = 500,
  aesthetics = "colour",
  breaks = ggplot2::waiver(),
  na.value = "grey50"
)
}
\description{
Automatically adjust color values
}
