% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-spruce.R
\name{scale_fill_spruce}
\alias{scale_fill_spruce}
\title{Automatically adjust fill values}
\usage{
scale_fill_spruce(
  ...,
  values,
  difference = 10,
  adjust = "lightness",
  filter = NULL,
  adjust_colors = NULL,
  range = NULL,
  maxit = 500,
  aesthetics = "fill",
  breaks = ggplot2::waiver(),
  na.value = "grey50"
)
}
\description{
Automatically adjust fill values
}
