% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-palette.R
\name{sort_colors}
\alias{sort_colors}
\title{Sort colors based on property}
\usage{
sort_colors(colors, property = "hue", desc = FALSE)
}
\arguments{
\item{colors}{Vector of colors}

\item{property}{Color properties to use for sorting}

\item{desc}{Sort in descending order}
}
\description{
Sort colors based on property
}
