% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce-colors.R
\name{spruce_colors}
\alias{spruce_colors}
\title{Adjust colors based on similarity}
\usage{
spruce_colors(
  colors,
  difference = 10,
  property = "interp",
  method = "CIE2000",
  range = NULL,
  filter = NULL,
  adjust_colors = NULL,
  exclude_colors = NULL,
  order = FALSE,
  maxit = 500,
  ...
)
}
\arguments{
\item{colors}{Character vector of colors to adjust}

\item{difference}{Color difference threshold (CIE200 score) to use for
adjusting `colors`.
Colors will be adjusted so the minimum pairwise difference
is greater than this threshold.}

\item{property}{Vector of color properties to adjust, can include any of:
- "interp", before adjusting color properties, first try interpolating new
  colors.
  This is the best approach for preserving the overall appearance of the
  original palette.
- "lightness", from LAB colorspace
- "a", from LAB colorspace
- "b", from LAB colorspace
- "hue", from HSL colorspace
- "saturation", from HSL colorspace
- "red", from RGB colorspace
- "green", from RGB colorspace
- "blue", from RGB colorspace}

\item{method}{Method to use for comparing colors, can be one of:
- "euclidian"
- "CIE1976"
- "CIE94"
- "CIE2000"
- "CMC"}

\item{range}{A vector containing the minimum and maximum values to use when
adjusting colors.
If multiple color properties will be adjusted, provide a named list with
ranges for each property.}

\item{filter}{Filter to apply to color palette when
calculating pairwise differences.
Colors will be adjusted to minimize the pairwise difference before and after
applying the filter.
A vector can be passed to adjust based on multiple color filters.
Possible values include,
- "colorblind", use deutan, protan, and tritan color blindness simulation
  filters
- "deutan"
- "protan"
- "tritan"}

\item{adjust_colors}{Index indicating color(s) to specifically adjust.
Should be an integer vector, or a character vector containing names matching
those provided for `colors`.}

\item{exclude_colors}{Index indicating color(s) to exclude when adjusting
palette.
Should be an integer vector, or a character vector containing names matching
those provided for `colors`.}

\item{order}{If `TRUE`, when `property` is "interp" colors will be ordered
based on the original palette.
If `FALSE`, adjusted colors will remain in their original positions.}

\item{maxit}{Maximum number of iterations to use when optimizing the color
palette.
Higher values will result in more optimal adjustments and a reduction in
speed.}

\item{...}{Additional control parameters to pass to `GenSA::GenSA()`.}
}
\description{
Adjust colors based on similarity
}
