% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-spruce.R
\name{spruce_up_colors}
\alias{spruce_up_colors}
\title{Adjust colors based on similarity}
\usage{
spruce_up_colors(
  colors,
  difference = 10,
  adjust = "lightness",
  range = NULL,
  filter = NULL,
  adjust_colors = NULL,
  exclude_colors = NULL,
  maxit = 500
)
}
\arguments{
\item{colors}{Character vector of colors to adjust}

\item{difference}{Color difference threshold (CIE200 score) to use for
adjusting `colors`.
Colors will be adjusted so the minimum pairwise difference
is greater than this threshold.}

\item{adjust}{Color attribute to adjust}

\item{range}{A vector containing the minimum and maximum values to use when
adjusting colors.}

\item{filter}{Filter to apply to color palette when
calculating pairwise differences.
Colors will be adjusted to minimize the pairwise difference before and after
applying the filter.
A vector can be passed to adjust based on multiple color filters.
Possible values include, "none", "deutan", "protan", and "tritan".}

\item{adjust_colors}{Index indicating color(s) to specifically adjust.
Should be an integer vector, or a character vector containing names matching
those provided for `colors`.}

\item{maxit}{Maximum number of iterations to use when optimizing the color
palette.
Higher values will result in more optimal adjustments and a reduction in
speed.}
}
\description{
Adjust colors based on similarity
}
