% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-colors.R
\name{subset_colors}
\alias{subset_colors}
\title{Subset colors based on pairwise differences or color properties}
\usage{
subset_colors(colors, ..., method = "CIE2000", filter = NULL)
}
\arguments{
\item{colors}{Vector of colors}

\item{...}{Subsetting specification, colors can be subsetted based on
minimum pairwise differences or color properties (refer to [get_property()]).
To subset based on color differences, use the `difference` variable,
e.g. `difference > 10`.
To subset based on color properties, refer to them by name,
e.g. `lightness > 50`.}

\item{method}{Method to use for comparing colors, refer to [compare_colors()]}

\item{filter}{Filter to apply to color palette when
calculating pairwise differences, refer to [compare_colors()]}
}
\description{
Subset colors based on pairwise differences or color properties
}
