% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-colors.R
\name{assign_colors}
\alias{assign_colors}
\title{Assign colors to provided names}
\usage{
assign_colors(colors, names, select_best = TRUE, order = TRUE, ...)
}
\arguments{
\item{colors}{Vector of colors}

\item{names}{Vector of names}

\item{select_best}{Select the most distinct colors when collapsing the
palette.
If `FALSE` colors will be selected starting with the first.}

\item{order}{If `TRUE`, when colors are expanded new colors are interspersed.
If `FALSE` new colors are added to the end.}

\item{...}{Additional arguments to pass to `resize_colors()`}
}
\description{
The color palette will be resized to accommodate the number of provided
names.
}
