% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-palette.R
\name{collapse_colors}
\alias{collapse_colors}
\title{Collapse color palette}
\usage{
collapse_colors(colors, n, filter = NULL)
}
\arguments{
\item{colors}{Vector of colors}

\item{n}{Number of colors to include in final color palette}

\item{filter}{Filter to apply to color palette when
calculating pairwise differences.
A vector can be passed to adjust based on multiple color filters.
Possible values include, "none", "deutan", "protan", and "tritan".}
}
\description{
Reduce the number of colors in palette by selecting the colors that
are most distinct
}
