% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce-colors.R
\name{compare_colors}
\alias{compare_colors}
\title{Compare colors in palette}
\usage{
compare_colors(colors, y = NULL, filter = NULL, method = "CIE2000")
}
\arguments{
\item{colors}{Vector of colors}

\item{y}{Vector of colors to compare,
if `NULL`, `colors` will be compared with itself.}

\item{filter}{Filter to apply to color palette when
calculating pairwise differences.
Colors will be adjusted to minimize the pairwise difference before and after
applying the filter.
A vector can be passed to adjust based on multiple color filters.
Possible values include, "deutan", "protan", and "tritan".}

\item{method}{Method to use for comparison,
refer to `farver::compare_colour()`.}
}
\description{
Compare colors in palette
}
