% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-palette.R
\name{expand_colors}
\alias{expand_colors}
\title{Expand color palette}
\usage{
expand_colors(colors, n, ...)
}
\arguments{
\item{colors}{Vector of colors}

\item{n}{Number of colors to include in final color palette}
}
\description{
Add additional colors to color palette
}
