% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize-palette.R
\name{resize_palette}
\alias{resize_palette}
\title{Resize color palette}
\usage{
resize_palette(colors, n, filter = NULL, ...)
}
\arguments{
\item{colors}{Vector of colors}

\item{n}{Number of colors to include in final color palette}

\item{filter}{Filter to apply to color palette when
calculating pairwise differences.
A vector can be passed to adjust based on multiple color filters.
Possible values include, "none", "deutan", "protan", and "tritan".}

\item{...}{Additional arguments to pass to `grDevices::colorRamp()`}
}
\description{
Reduce or expand the number of colors in palette
}
