% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-colors.R
\name{sort_colors}
\alias{sort_colors}
\title{Sort colors based on property}
\usage{
sort_colors(colors, ..., method = "CIE2000", filter = NULL, desc = FALSE)
}
\arguments{
\item{colors}{Vector of colors}

\item{...}{Color attribute(s) to use for sorting, this can include
minimum pairwise color differences or color properties
(refer to [get_property()]).
Specify 'difference' to sort based on color differences.}

\item{method}{Method to use for comparing colors, refer to [compare_colors()]}

\item{filter}{Filter to apply to color palette when
calculating pairwise differences, refer to [compare_colors()]}

\item{desc}{Sort in descending order}
}
\description{
Sort colors based on property
}
